pyproject.toml
[tool.poetry]
name = "homicide-bot"
version = "0.1.0"
description = "Tweeting information about yesterday's homicides in Philadelphia"
authors = ["Nick Hand <nicholas.adam.hand@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.1"
bs4 = "^0.0.1"
tweepy = "^4.4.0"
pandas = "^1.3.4"
requests = "^2.26.0"
click = "^8.0.3"
python-dotenv = "^0.19.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
homicide-bot = "homicide_bot.__main__:main"